+++ plugins/ci/pkg/commands/commands.go
@@ -14,8 +14,9 @@ func ExecInDir(dir string, cmd *exec.Cmd, message string) (string, error) {
 	output, err := cmd.CombinedOutput()
 	if err != nil {
 		logrus.Errorf("Error running %s - %s[%v]", util.RemoveToken(cmd.String()), string(output), util.RemoveToken(err.Error()))
+		return "", err
 	}
-	return string(output), err
+	return string(output), nil
 }
 
 // ExecWithMessage runs a command and prints errors to Stderr
@@ -24,16 +25,18 @@ func ExecWithMessage(cmd *exec.Cmd, message string) (string, error) {
 	output, err := cmd.CombinedOutput()
 	if err != nil {
 		logrus.Errorf("Error running %s - %s[%v]", util.RemoveToken(cmd.String()), string(output), util.RemoveToken(err.Error()))
+		return "", err
 	}
-	return string(output), err
+	return string(output), nil
 }
 
 // Exec executes a command and returns the results as a string.
 func Exec(command string, args ...string) (string, error) {
 	cmd := exec.Command(command, args...)
-	bytes, err := cmd.CombinedOutput()
+	output, err := cmd.CombinedOutput()
 	if err != nil {
-		logrus.Errorf("Unable to execute command: %s %s[%v]", util.RemoveToken(cmd.String()), string(bytes), util.RemoveToken(err.Error()))
+		logrus.Errorf("Unable to execute command: %s %s[%v]", util.RemoveToken(cmd.String()), string(output), util.RemoveToken(err.Error()))
+		return "", err
 	}
-	return string(bytes), err
+	return string(output), nil
 }